Beginning of refactor.  BREAKING: default format is now "guess".
