--- .npmignore
@@ -1,10 +1,0 @@
-.c8rc
-.github/
-.ncurc
-.vscode/
-coverage/
-eslint.config.js
-pnpm-lock.yaml
-test/
-tsconfig.json
-eslint.config.mjs
