+++ lib/index.js
@@ -9,72 +9,10 @@
 const { Module } = require("node:module");
 const fs = require("node:fs/promises");
 const path = require("node:path");
-const semver = require("semver");
-const url = require("node:url");
 const vm = require("node:vm");
-
-// These already exist in a new, blank VM.  Date, JSON, NaN, etc.
-// Things from the core language.
-const vmGlobals = new Set(new vm
-  .Script("Object.getOwnPropertyNames(globalThis)")
-  .runInNewContext());
-vmGlobals.add("global");
-vmGlobals.add("globalThis");
-vmGlobals.add("sys");
-
-// These are the things that are normally in the environment, that vm doesn't
-// make available.  This that you expect to be available in a node environment
-// that aren't in the laguage itself.  There are a lot more things in this list
-// than you expect, like setTimeout and structuredClone.
-const neededKeys = Object
-  .getOwnPropertyNames(global)
-  .filter(k => !vmGlobals.has(k))
-  .sort();
-const globalContext = Object.fromEntries(
-  neededKeys.map(k => [k, global[
-    /** @type {keyof typeof global} */ (k)
-  ]])
-);
-
-globalContext.console = console;
-
-/**
- * @typedef {"amd"
- * | "bare"
- * | "cjs"
- * | "commonjs"
- * | "es"
- * | "es6"
- * | "esm"
- * | "globals"
- * | "guess"
- * | "mjs"
- * | "module"
- * | "umd" } SourceFormat
- */
-
-/**
- * Options for how to process code.
- *
- * @typedef {object} FromMemOptions
- * @property {SourceFormat} [format="commonjs"] What format does the code
- *   have?  "guess" means to read the closest package.json file looking for
- *   the "type" key.  "globals", "amd", and "bare" are not actually supported.
- * @property {Record<string, any>} [env] If specified, use this instead of the
- *   current values in process.env.  Works if includeGlobals is false by
- *   creating an otherwise-empty process instance.
- * @property {string} filename What is the fully-qualified synthetic filename
- *   for the code?  Most important is the directory, which is used to find
- *   modules that the code import's or require's.
- * @property {Record<string, any>} [context={}] Variables to make availble in
- *   the global scope while code is being evaluated.
- * @property {boolean} [includeGlobals=true] Include the typical global
- *   properties that node gives to all modules.  (e.g. Buffer, process).
- * @property {number} [lineOffset=0] Specifies the line number offset that is
- *   displayed in stack traces produced by this script.
- * @property {number} [columnOffset=0] Specifies the first-line column number
- *   offset that is displayed in stack traces produced by this script.
- */
+const globalContext = require("./global.js");
+const isImportSupported = require("./supported.js");
+const importString = require("./importString.js");
 
 /**
  * Treat the given code as a node module as if require() had been called
@@ -103,91 +41,6 @@ function requireString(code, dirname, options) {
     __filename: options.filename,
     ...options.context,
   });
-}
-
-/**
- * If the given specifier starts with a ".", path.resolve it to the given
- * directory.  Otherwise, it's a fully-qualified path, a node internal
- * module name, an npm-provided module name, or a URL.
- *
- * @param {string} dirname Owning directory
- * @param {string} specifier String from the rightmost side of an import statement
- * @returns {string} Resolved path name or original string
- */
-function resolveIfNeeded(dirname, specifier) {
-  if (specifier.startsWith(".")) {
-    specifier = new URL(specifier, dirname).toString();
-  }
-  return specifier;
-}
-
-/**
- * Is importing from an ES6 module supported in the current environment?
- *
- * @returns {boolean}
- */
-function isImportSupported() {
-  return Boolean(vm.SourceTextModule);
-}
-
-/**
- * Treat the given code as a node module as if import had been called
- * on a file containing the code.
- *
- * @param {string} code Source code in es6 format.
- * @param {string} dirname Where the synthetic file would have lived.
- * @param {FromMemOptions} options
- * @returns {Promise<unknown>} The module exports from code
- */
-async function importString(code, dirname, options) {
-  if (!isImportSupported()) {
-    throw new Error("Start node with `node --experimental-vm-modules` in order to import ES6 modules");
-  }
-
-  if (!semver.satisfies(process.version, ">=20.8")) {
-    throw new Error("Requires node.js 20.8+ or 21.");
-  }
-
-  const fileUrl = options.filename.startsWith("file:")
-    ? options.filename
-    : url.pathToFileURL(options.filename).toString();
-  const dirUrl = dirname.startsWith("file:")
-    ? dirname + "/"
-    : url.pathToFileURL(dirname).toString() + "/";
-
-  const mod = new vm.SourceTextModule(code, {
-    identifier: fileUrl,
-    lineOffset: options.lineOffset,
-    columnOffset: options.columnOffset,
-    context: vm.createContext(options.context),
-    initializeImportMeta(meta) {
-      meta.url = fileUrl;
-    },
-    importModuleDynamically(specifier) {
-      return import(
-        /* webpackIgnore: true */ resolveIfNeeded(dirUrl, specifier)
-      );
-    },
-  });
-
-  await mod.link(async(specifier, referencingModule) => {
-    const resolvedSpecifier = resolveIfNeeded(dirUrl, specifier);
-    const targetModule = await import(
-      /* webpackIgnore: true */ resolvedSpecifier
-    );
-    const exports = Object.keys(targetModule);
-
-    // DO NOT change function to () =>, or `this` will be wrong.
-    return new vm.SyntheticModule(exports, function() {
-      for (const e of exports) {
-        this.setExport(e, targetModule[e]);
-      }
-    }, {
-      context: referencingModule.context,
-    });
-  });
-  await mod.evaluate();
-  return mod.namespace;
 }
 
 /**
@@ -261,7 +114,7 @@ guessModuleType.clearCache = function clearCache() {
  */
 async function fromMem(code, options) {
   options = {
-    format: "commonjs",
+    format: "guess",
     env: undefined,
     includeGlobals: true,
     lineOffset: 0,
